refactor(swap): simplify swap argument handling by removing redundant functions
- Removed `swapArgsCallback` and `swapPayableArgsCallback` functions to streamline the codebase.
- Updated imports in `index.ts` to reflect the removal of the deprecated functions.
- This change enhances code readability and maintainability by consolidating argument preparation logic.
